Add SandboxDebugEnv for sandbox tasksets
